update revert tests, add solcover
