+++ .solcover.js
@@ -1,0 +1,3 @@
+module.exports = {
+  skipFiles: ['./mock'],
+}
