+++ test/WithdrawFee.ts
@@ -56,7 +56,7 @@ describe("Withdraw fees", function () {
         const partyLP = await v3PartyFactory.getPool(await weth9.getAddress(), MEME, FeeAmount.HIGH)
         await bnbPartyFactory.withdrawPartyLPFee([partyLP])
         const balanceAfter = await weth9.balanceOf(await signers[0].getAddress())
-        expect(balanceAfter).to.be.equal(balanceBefore + expectedFee - 1n)
+        expect(balanceAfter).to.be.equal(balanceBefore + expectedFee)
     })
 
     it("should return zero if pool is zero address", async () => {
@@ -105,13 +105,6 @@ describe("Withdraw fees", function () {
 
     it("isToken0WBNB should return true if token0 is WBNB", async () => {
         expect(await bnbPartyFactory.isToken0WBNB(lpAddress)).to.be.true
-    })
-
-    it("isToken0WBNB should revert if set zero address", async () => {
-        await expect(bnbPartyFactory.isToken0WBNB(ethers.ZeroAddress)).to.be.revertedWithCustomError(
-            bnbPartyFactory,
-            "ZeroAddress"
-        )
     })
 
     it("should deacrease fee after withdraw", async () => {
